+++ NEWS.md
@@ -3,4 +3,8 @@
 * Added a `NEWS.md` file to track changes to the package.
 * Added `ep_build()` and initial subfunctions to preprocess, predict, and summarise EP
 * Added `ep_load_*()` to download release data
-* Added `update/` script that uses piggyback to update data locally
+* Added `update/` script that uses piggyback to update release data
+* Offloaded models to releases ("latest-models", "v1.0.0-models")
+* Added download of models to user cache
+* Combine variants of ep_load into one function `ep_load()` with arguments
+